Moved templates & webview manager into it's own submodule
